+++ rplugin/python/ensime.py
@@ -7,6 +7,7 @@ import re
 import base64
 import logging
 import time
+import datetime
 import thread
 import inspect
 from socket import error as socket_error
@@ -17,6 +18,8 @@ import Queue
 class EnsimeLauncher:
     def __init__(self, conf_path):
         self.generating_classpath = False
+        self.process = None
+        self.log_file = None
         self.classpath = None
         self.conf_path = conf_path
         self.version = "0.9.10-SNAPSHOT"
@@ -71,12 +74,11 @@ saveClasspathTask := {
             self.write_file("{}/build.sbt".format(self.classpath_dir), build_sbt)
             self.write_file("{}/project/build.properties".format(self.classpath_dir),
                     "sbt.version=0.13.8")
+            self.log_file = open('{}/saveClasspath.log'.format(
+                self.classpath_dir), 'w')
             cwd = os.getcwd()
             os.chdir(self.classpath_dir)
-            log_file = open('saveClasspath.log', 'w')
-            self.process = subprocess.Popen(["sbt", "saveClasspath"],
-                    stdout=log_file,
-                    stderr=subprocess.STDOUT)
+            self.process = os.spawnl(os.P_WAIT, "sbt", "/usr/bin/sbt", "saveClasspath")
             os.chdir(cwd)
     def read_file(self, path):
         f = open(path)
@@ -107,7 +109,7 @@ saveClasspathTask := {
                         self.read_file(self.classpath_file),
                         self.conf['java-home'])
     def run(self):
-        if self.conf != None and not self.is_running():
+        if self.classpath != None and self.conf != None and not self.is_running():
             if not os.path.exists(self.conf['cache-dir']):
                 os.mkdir(self.conf['cache-dir'])
             self.log_file = open(self.conf_path + '_cache/server.log', 'w')
@@ -124,8 +126,8 @@ saveClasspathTask := {
     def wait(self):
         self.process.wait()
     def stop(self):
-        os.kill(self.process.pid, signal.SIGTERM)
-        self.log_file.close()
+        if self.process != None: os.kill(self.process.pid, signal.SIGTERM)
+        if self.log_file != None: self.log_file.close()
 @neovim.plugin
 class Ensime(object):
     def log(self, what):
@@ -133,7 +135,7 @@ class Ensime(object):
         if os.path.isdir(self.ensime_cache):
             log_dir = self.ensime_cache
         f = open(log_dir + "ensime-vim.log", "a")
-        f.write("{}: {}\n".format(time.strftime("%Y-%m-%d %H:%M:%S"), what))
+        f.write("{}: {}\n".format(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f"), what))
         f.close()
     def unqueue_poll(self):
         while True:
@@ -166,13 +168,18 @@ class Ensime(object):
     def start_ensime_launcher(self):
         if self.ensime == None:
             self.ensime = EnsimeLauncher(".ensime")
+            self.log("EnsimeLauncher instantiation")
+            self.message("ensime setup, generating classpath " 
+                    + "may take a while the first time...")
         if self.ensime.classpath != None:
+            self.log("starting up ensime")
             self.message("ensime startup")
             self.ensime.run()
             return True
         else:
-            self.message("ensime setup, generating classpath may take a while the first time...")
-            self.ensime.setup()
+           self.log("launching EnsimeLauncher.setup()")
+           self.ensime.setup()
+           self.log("done launching EnsimeLauncher.setup()")
         return False
     def stop_ensime_launcher(self):
         self.ensime.stop()
